refactor: Move edit logic into separate sub-app
This follows the pattern we started to use for the move journeys to
group the logic within separate self-contained sub-apps.

This should improve maintainance and reduce the size of some of the
existing folders.
