--- app/move/config/index.js
@@ -1,5 +1,0 @@
-const updateConfig = require('./update.config')
-
-module.exports = {
-  updateConfig,
-}
