+++ app/move/controllers/index.js
@@ -1,11 +1,9 @@
 const confirmation = require('./confirmation')
 const timeline = require('./timeline')
-const update = require('./update')
 const view = require('./view')
 
 module.exports = {
   confirmation,
   timeline,
-  update,
   view,
 }
