--- app/move/steps/index.js
@@ -1,5 +1,0 @@
-const update = require('./update')
-
-module.exports = {
-  update,
-}
